\name{load_commit_stats}
\alias{load_commit_stats}
\title{Loads commit stats (as generated by cloc) into a data frame, along with the commit ref and timestamp.}
\usage{
  load_commit_stats(path)
}
\arguments{
  \item{path}{the path to the commit line count statistics.
  See linecount.sh}
}
\value{
  data.frame of all commit line count stats
}
\description{
  Loads commit stats (as generated by cloc) into a data
  frame, along with the commit ref and timestamp.
}
\examples{
commit_stats <- load_commit_stats()
summary(commit_stats)
}

