\name{plot_commit_stats}
\alias{plot_commit_stats}
\title{Plots the variable for each language at each commit time.}
\usage{
  plot_commit_stats(commit_stats, variable, title)
}
\arguments{
  \item{commit_stats}{}

  \item{variable}{the variable to plot}

  \item{title}{the plot title}
}
\value{
  ggplot object
}
\description{
  Plots the variable for each language at each commit time.
}

