\name{plot_commit_stats_line}
\alias{plot_commit_stats_line}
\title{Plots the variable as a line chart vs. Time + languages.}
\usage{
  plot_commit_stats_line(commit_stats, variable, title)
}
\arguments{
  \item{commit_stats}{}

  \item{variable}{the variable to plot}

  \item{title}{the plot title}
}
\value{
  ggplot object
}
\description{
  Plots the variable as a line chart vs. Time + languages.
}

